ref: move table components to table folder
