+++ src/components/table/VueScreenerTableRow.vue
@@ -6,19 +6,18 @@
 
 <script lang="ts" setup>
 import { computed } from 'vue'
-import { twMerge } from '../../../utils/tailwind-merge.utils'
+import { twMerge } from '../../utils/tailwind-merge.utils'
 
-export type TableHeadUI = {
+export type TableRowUI = {
   class?: string
 }
 
 const props = defineProps<{
-  ui?: TableHeadUI
+  ui?: TableRowUI
 }>()
 
 const uiDefaults = {
-  class:
-    'vsc-grid vsc-col-span-full vsc-grid-cols-subgrid vsc-font-bold vsc-border-b vsc-border-zinc-700 vsc-sticky vsc-top-0 vsc-h-fit',
+  class: 'vsc-grid vsc-col-span-full vsc-grid-cols-subgrid vsc-border-b vsc-border-zinc-700 vsc-h-fit', // eslint-disable-line
 }
 
 const ui = computed(() => {
