+++ src/components/views/TableView.vue
@@ -79,10 +79,10 @@
 <script lang="ts" setup>
 import { computed } from 'vue'
 import type { VueScreener, ColDef } from '../../interfaces/screener'
-import VueScreenerTableRow, { TableRowUI } from '../ui/table/VueScreenerTableRow.vue'
-import VueScreenerTableCell, { TableCellUI } from '../ui/table/VueScreenerTableCell.vue'
-import VueScreenerTable, { TableUI } from '../ui/table/VueScreenerTable.vue'
-import VueScreenerTableHead, { TableHeadUI } from '../ui/table/VueScreenerTableHead.vue'
+import VueScreenerTableRow, { TableRowUI } from '../table/VueScreenerTableRow.vue'
+import VueScreenerTableCell, { TableCellUI } from '../table/VueScreenerTableCell.vue'
+import VueScreenerTable, { TableUI } from '../table/VueScreenerTable.vue'
+import VueScreenerTableHead, { TableHeadUI } from '../table/VueScreenerTableHead.vue'
 import SortIcon, { SortIconUI } from '../icons/SortIcon.vue'
 import { highlightMatches } from '../../utils/text.utils'
 import { twMerge } from '../../utils/tailwind-merge.utils'
