+++ src/index.ts
@@ -2,7 +2,7 @@ import VueScreener from './components/VueScreener.vue'
 import VueScreenerPagination from './components/pagination/VueScreenerPagination.vue'
 import VueScreenerSearch from './components/search/VueScreenerSearch.vue'
 import VueScreenerSearchOptions from './components/search/VueScreenerSearchOptions.vue'
-import VueScreenerTableCell from './components/ui/table/VueScreenerTableCell.vue'
+import VueScreenerTableCell from './components/table/VueScreenerTableCell.vue'
 import { useVueScreener } from './hooks/use-vue-screener'
 import './assets/tailwind.css'
